[CCUBE-2188][GZ] extract styling to component style file
