+++ src/file-upload/file-list-item/file-list-item.styles.ts
@@ -3,7 +3,6 @@ import { css } from "@linaria/core";
 import {
     Border,
     Colour,
-    Font,
     MediaQuery,
     Radius,
     Shadow,
@@ -65,41 +64,6 @@ export const editFileNameText = css`
 
 export const editFileSizeText = css`
     color: ${Colour["text-subtler"]};
-`;
-
-export const actionButtonsSection = css`
-    display: flex;
-    justify-content: flex-end;
-
-    ${MediaQuery.MaxWidth.md} {
-        flex-direction: column;
-    }
-`;
-
-export const actionButtonsSectionWithThumbnail = css`
-    margin-left: calc(
-        96px + ${Spacing["spacing-32"]}
-    ); // thumbnail width + right margin
-
-    ${MediaQuery.MaxWidth.md} {
-        margin-left: 0;
-    }
-`;
-
-export const actionButton = css`
-    width: 7.5rem;
-
-    &:not(:last-of-type) {
-        margin-right: ${Spacing["spacing-16"]};
-    }
-
-    ${MediaQuery.MaxWidth.md} {
-        width: 100%;
-
-        &:not(:last-of-type) {
-            margin-bottom: ${Spacing["spacing-16"]};
-        }
-    }
 `;
 
 // =============================================================================
@@ -206,98 +170,5 @@ export const contentSection = css`
             flex-direction: row;
             align-items: center;
         }
-    }
-`;
-
-export const nameSection = css`
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    width: 100%;
-`;
-
-export const extendedNameSection = css`
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    align-items: flex-start;
-
-    ${MediaQuery.MaxWidth.md} {
-        width: 100%;
-    }
-`;
-
-export const descriptionFileSizeText = css`
-    color: ${Colour["text-subtler"]};
-    margin-top: ${Spacing["spacing-16"]};
-`;
-
-export const itemDescriptionLabel = css`
-    ${Font["form-label"]}
-    font-weight: ${Font.Spec["weight-semibold"]};
-    color: ${Colour["text-subtle"]};
-    margin-top: ${Spacing["spacing-16"]};
-`;
-
-export const itemDescriptionText = css`
-    margin-top: ${Spacing["spacing-4"]};
-    overflow-wrap: anywhere;
-`;
-
-export const errorIcon = css`
-    height: 1lh;
-    width: 1em;
-    flex-shrink: 0;
-    color: ${Colour["icon-error-strong"]};
-`;
-
-export const errorMessage = css`
-    color: ${Colour["text-error"]};
-    margin-top: ${Spacing["spacing-4"]};
-    display: flex;
-    gap: ${Spacing["spacing-4"]};
-`;
-
-export const actionContainer = css`
-    width: 6rem;
-    margin-left: ${Spacing["spacing-32"]};
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-
-    ${MediaQuery.MaxWidth.md} {
-        width: fit-content;
-    }
-
-    &[data-mobile-layout="loading"] {
-        ${MediaQuery.MaxWidth.md} {
-            margin-left: 0;
-            margin-top: ${Spacing["spacing-16"]};
-            width: 100%;
-        }
-    }
-
-    &[data-mobile-layout="editable"] {
-        ${MediaQuery.MaxWidth.md} {
-            margin-left: 0;
-            margin-top: ${Spacing["spacing-16"]};
-            align-self: flex-end;
-        }
-    }
-`;
-
-export const inlineActionContainer = css`
-    width: 100%;
-    margin-top: ${Spacing["spacing-16"]};
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-`;
-
-export const iconButton = css`
-    min-width: unset;
-
-    &:not(:last-child) {
-        margin-right: ${Spacing["spacing-8"]};
     }
 `;
